+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/). All notable changes will be documented in this file.
 
+## [Unreleased](https://github.com/Boreal-Info-Sys/borealis-pg-cli/compare/v1.7.1...HEAD)
+- Provide data integration/PostgreSQL extension name hint in removal command prompts
+
 ## [1.7.1](https://github.com/Boreal-Info-Sys/borealis-pg-cli/compare/v1.7.0...v1.7.1)
 - Fixed: Handle errors for commands that attempt DDL during a PostgreSQL major version upgrade